requirements.txt
attrs==22.2.0
click==8.1.3
exceptiongroup==1.1.0
flake8==6.0.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.1
python-dotenv==0.21.1
SQLAlchemy==2.0.3
tomli==2.0.1
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.13.0
